+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
 
-#### 9th December 2020 - Version 3.5.1
+#### 11th December 2020 - Version 3.5.2
 
 -	Tweak: Added Tiktok icon
-
+-   Fix: Advanced slider bug with WP 5.6
 #### 6th November 2020 - Version 3.5.0
 
 -	Tweak: Do not output span shortcode ID attribute if empty